Dialogue system update
I added a new dialogue type called narration. I did this so I could use the system to display the pop-up I wanted to show for the end of the game.
I also added some regions and comments to the system.

Also updated the prefab to have a narration dialogue box.