Bed now transitions back to main menu
It's not a smooth transition, I was trying to add a way to have smooth transitions between scenes.
I spent a while fiddling with that cause I wasn't super familiar with the scene manager. As that implies, I was learning how scenes work while I was doing this. And I've realised now that we'd need to re-structure how we organise scenes and menus in the game to have smooth transitions.
We'd have to make a fairly complex system to make it work properly.

But that's me done for now, I'm not looking to make that system any time soon.