Sorting Layer changes
Infront layer removed
And the HUD spriterenderes layer too as it is un-used
The reason Infront had to be removed is due to some weirdness with how the blur works. The blur requires the CameraSortingLayer to exist and this layer interferes with the layers below it.
For some reason it was making everything in infront not display.

Changes have been made to the main scene in this branch to move all the sprites that were using that layer to default layer. I just set their number to the maximum value in the default layer, they are displaying the same as before.